Safari-611.1.10/LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/datetime-local-expected.txt


PASS empty value
PASS datetime-local input value set to 2014-01-01T11:11:11.111 without min/max
FAIL datetime-local input value set to 2014-01-01 11:11:11.111 without min/max assert_equals: expected "2014-01-01T11:11:11.111" but got ""
FAIL datetime-local input value set to 2014-01-01 11:11 without min/max assert_equals: expected "2014-01-01T11:11" but got ""
FAIL datetime-local input value set to 2014-01-01 00:00:00.000 without min/max assert_equals: expected "2014-01-01T00:00" but got ""
PASS datetime-local input value set to 2014-01-0 11:11 without min/max
PASS datetime-local input value set to 2014-01-01 11:1 without min/max
PASS invalid datetime-local input value 1
PASS invalid datetime-local input value 2
PASS invalid datetime-local input value 3
PASS invalid datetime-local input value 4
PASS invalid datetime-local input value 5
PASS invalid datetime-local input value 6
FAIL Value >= min attribute assert_equals: expected "2014-01-01T11:12" but got ""
FAIL Value < min attribute assert_equals: expected "2014-01-01T11:11" but got ""
FAIL Value <= max attribute assert_equals: expected "2014-01-01T11:10" but got ""
FAIL Value > max attribute assert_equals: expected "2014-01-01T11:11" but got ""
